requirements.txt
aiogram
opencv-python
aiohttp
python-dotenv
dlib
face_recognition
numpy
